chore: Rename `InsightsView.ts` to just `View.ts`
