+++ mobile/src/modules/insights/screens/View.tsx
